wrap eject around prompt
