+++ bin/eject-react-app.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+require('../scripts/eject');
