--- bin/graduate-react-app.js
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-require('../scripts/graduate');
