+++ package.json
@@ -20,7 +20,7 @@
   "bin": {
     "start-react-app": "./bin/start-react-app.js",
     "build-react-app": "./bin/build-react-app.js",
-    "graduate-react-app": "./bin/graduate-react-app.js"
+    "eject-react-app": "./bin/eject-react-app.js"
   },
   "dependencies": {
     "autoprefixer": "^6.3.7",
